Move format methods to AbstractPGXS and ...
obtain flags from pg_config
